requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
redis==5.0.1
python-dotenv==1.0.0
pydantic==2.6.1
langchain-core==0.1.18
langchain-openai==0.0.5
langgraph==0.0.19
langchain-redis==0.0.5 